Renamed types.py to registries.py
Pulled all internal registry decorators for types and schemas into
_registered_types.py.

Reorganized and refactored some of the types to be consistent
