+++ datagen/preprocessor.py
@@ -7,12 +7,12 @@ import json
 import logging
 from urllib.parse import parse_qs
 from .exceptions import SpecException
-from . import types
+from . import registries
 
 _log = logging.getLogger(__name__)
 
 
-@types.registry.preprocessors('default')
+@registries.registry.preprocessors('default')
 def _preprocess_spec(raw_spec):
     """
     Preprocesses the spec into a format that is easier to use.
@@ -40,7 +40,7 @@ def _preprocess_spec(raw_spec):
     return updated_specs
 
 
-@types.registry.preprocessors('csv-select')
+@registries.registry.preprocessors('csv-select')
 def _preprocess_csv_select(raw_spec):
     """
     Converts and csv-select elements into standard csv ones
@@ -81,7 +81,7 @@ def _preprocess_csv_select(raw_spec):
     return updated_specs
 
 
-@types.registry.preprocessors('nested')
+@registries.registry.preprocessors('nested')
 def _preprocess_nested(raw_spec):
     """
     Converts all nested elements
@@ -125,7 +125,7 @@ def _preprocess_nested(raw_spec):
     return updated_specs
 
 
-@types.registry.preprocessors('type_check')
+@registries.registry.preprocessors('type_check')
 def _preprocess_verify_types(raw_spec):
     """ log only checks """
     for key, field_spec in raw_spec.items():
@@ -133,9 +133,9 @@ def _preprocess_verify_types(raw_spec):
             _preprocess_verify_types(field_spec)
             continue
         type_name = field_spec.get('type')
-        if types.lookup_type(type_name) is None:
+        if registries.lookup_type(type_name) is None:
             _log.warning('Unknown type key: %s for spec %s, known types are %s',
-                         type_name, field_spec, types.registered_types())
+                         type_name, field_spec, registries.registered_types())
     return raw_spec
