+++ tests/test_range_supplier.py
@@ -1,59 +1,8 @@
 import pytest
 
 from datagen import builder, Loader, SpecException
-from datagen import range_suppliers
-
-
-def test_range_missing_data():
-    with pytest.raises(SpecException):
-        range_suppliers._configure_range_supplier({'type': 'range'}, None)
-
-
-def test_range_invalid_data_type():
-    with pytest.raises(SpecException):
-        range_suppliers._configure_range_supplier({'type': 'range', 'data': 42}, None)
-
-
-def test_range_end_before_start():
-    start = 10
-    end = 9
-    with pytest.raises(SpecException):
-        range_suppliers._configure_range_supplier({'type': 'range', 'data': [start, end]}, None)
-
-
-def test_range_valid():
-    start = 2
-    end = 10
-    step = 2
-    supplier = range_suppliers._configure_range_supplier({'type': 'range', 'data': [start, end, step]}, None)
-
-    expected = [2, 4, 6, 8, 10]
-    actual = [supplier.next(i) for i in range(5)]
-
-    assert expected == actual
-
-
-def test_range_float_step():
-    start = 0
-    end = 1
-    step = .1
-    supplier = range_suppliers._configure_range_supplier({'type': 'range', 'data': [start, end, step]}, None)
-
-    expected = [0, .1, .2, .3, .4, .5, .6, .7, .8, .9, 1]
-    actual = [supplier.next(i) for i in range(11)]
-
-    assert expected == actual
-
-
-def test_range_float_start_end():
-    start = 0.5
-    end = 5.5
-    supplier = range_suppliers._configure_range_supplier({'type': 'range', 'data': [start, end]}, None)
-
-    expected = [0.5, 1.5, 2.5, 3.5, 4.5, 5.5]
-    actual = [supplier.next(i) for i in range(6)]
-
-    assert expected == actual
+# to trigger registration
+from datagen import cli
 
 
 def test_rand_range():
@@ -107,18 +56,3 @@ def test_nested_range_lists_mixed_types_and_step_cast():
 
     assert supplier.next(0) == '0.5'
     assert supplier.next(1) == '20.01'
-
-
-def test_float_range1():
-    range_list = list(range_suppliers._float_range(1.0, 5.0, 1.0))
-    assert range_list == [1.0, 2.0, 3.0, 4.0]
-
-
-def test_float_range_with_precision1():
-    range_list = list(range_suppliers._float_range(1.25, 2.25, 0.25, 2))
-    assert range_list == [1.25, 1.5, 1.75, 2.0]
-
-
-def test_float_range_with_precision2():
-    range_list = list(range_suppliers._float_range(1.2499999, 2.2499999, 0.25, 2))
-    assert range_list == [1.25, 1.5, 1.75, 2.0]
