+++ tests/test_spec_formatters.py
@@ -1,5 +1,3 @@
-import pytest
-
 import datagen
