+++ tests/test_uuid.py
@@ -3,8 +3,8 @@ import re
 import pytest
 
 from datagen import builder, Loader
-# need this to trigger registration
-from datagen import uuid_handler
+# to trigger registration
+from datagen import cli
 
 
 UUID_REGEX = re.compile('[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}', re.I)
