+++ tests/test_yaml_specs.py
@@ -3,7 +3,7 @@ import json
 
 from datagen.loader import Loader
 # to trigger registration
-from datagen import combine
+from datagen import cli
 
 
 yaml_spec = '''
